+++ RELEASE_NOTES.md
@@ -1,6 +1,11 @@
 # kb_motupan release notes
 =========================================
 
+0.0.1
+-----
+* Added mmseqs2 and mOTUlizer installs to Dockerfile
+* Added run_mmseqs2_and_motupan_files() method
+
 0.0.0
 -----
 * Module created by kb-sdk init